SO MANY DATA TABLES
* Added a bunch of data tables
* Changed feat focus generator, so that if you give it a focus type that
is actually just a focus, it just returns that. Means I have to add new
focus type groups when I know of them, because no tests will fail to
tell me I need to add it.
* Only get required feat Ids if it has required feats (means I can't
forget to add them to the table, or no tests will warning me)
* Corrected the weapon type of whips and hand crossbows for bards and
rogues
* Added count verification of collections in case no extras appear
because of duplicates