Fix Weapon generator returning null
Occasionally, it couldn't make an appropriate weapon in 10,000 tries.
Set it up to try again with mundane, and if still failing, give them a
club (everyone can wield a club).

It sucks that this now has to know the proper way to build a club (only
the treasure generator should be concerned with that), but it's not too
terrible.

* Also got rid of unneeded stress tests (testing curses and
intelligence, which are completely handled by the treasure generator)
* Made ranged weapon stress test also check for melee weapon